# swagger.yaml
openapi: 3.0.3
info:
  title: SET Service API
  version: 1.1.0
  description: |
    SET Service REST API for Worker, Company, Super Admin, and restricted Admin roles.

    Core worker flow:
    Worker registers -> verifies OTP -> becomes pending_approval -> super admin or permitted admin approves or rejects -> approved worker can log in and receive tokens.

    Contract-wide observability and authentication rules:
    - Every HTTP response includes `X-Request-ID` and `X-Correlation-ID`.
    - Clients may send either header using 8-128 characters from `[A-Za-z0-9._:-]`; invalid values are replaced server-side.
    - Error bodies include the effective `request_id`; the correlation identifier remains in the response header.
    - `BearerAuth` accepts access tokens only. Refresh tokens are valid only in the documented refresh/logout request bodies.
    - OTP request success means the request was accepted for delivery; it does not prove arrival at the handset or mailbox.
    - Interactive Swagger UI is disabled in production and should be exposed only in controlled non-production environments.
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.setservice.az/v1
    description: Production
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT access token
      description: Short-lived SET Service JWT access token. Refresh tokens are rejected by authenticated API middleware.
  headers:
    XRequestId:
      description: Effective request identifier used in structured logs and error responses.
      schema: { $ref: '#/components/schemas/RequestId' }
    XCorrelationId:
      description: Correlation identifier propagated across related requests; defaults to the request identifier.
      schema: { $ref: '#/components/schemas/RequestId' }
...
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
    RequestId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      example: 7fa22d18-6e54-4c33-8154-9f68f8e2f84b
    Error:
      type: object
      required: [error, code, request_id, timestamp]
      properties:
        error: { type: string, example: Worker is not approved for login. }
        code: { type: string, example: WORKER_NOT_APPROVED }
        details: { type: object, nullable: true, example: { status: pending_approval } }
        request_id: { $ref: '#/components/schemas/RequestId' }
        timestamp: { type: string, format: date-time }
...
    WorkerStatus:
      type: string
      enum: [draft, pending_otp, pending_approval, approved, rejected, suspended, inactive]
      example: pending_approval
    WorkerClass:
      type: string
      nullable: true
      enum: [A, B, C]
      example: A
...
    Skill:
      oneOf:
        - type: string
          example: waiter
        - type: object
          properties:
            name: { type: string, example: waiter }
            level: { type: integer, minimum: 1, maximum: 5, example: 4 }
...
    WorkerDocumentDeletionResponse:
      type: object
      required: [status, worker_id, document_type, deleted_at]
      properties:
        status: { type: string, enum: [deleted] }
        worker_id: { type: string, format: uuid }
        document_type: { type: string, enum: [health_certificate, criminal_record] }
        deleted_at: { type: string, format: date-time }
    WorkerAccountDeletionRequest:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm:
          type: boolean
          enum: [true]
          description: Explicit confirmation that the authenticated worker wants the account deletion request applied.
    WorkerAccountDeletionResponse:
      type: object
      required: [request_id, worker_id, status, account_state, effective_at]
      properties:
        request_id: { type: string, format: uuid }
        worker_id: { type: string, format: uuid }
        status: { type: string, enum: [accepted] }
        account_state: { type: string, enum: [inactive] }
        effective_at: { type: string, format: date-time }
    WorkExperience:
      type: object
      required: [company_name, position]
      properties:
        company_name: { type: string, example: Marriott Boulevard }
        position: { type: string, example: Ofisiant }
        note: { type: string, example: Banket və VIP qonaq xidməti }
...
    WorkerProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        email: { type: string, format: email, nullable: true, description: Visible to worker and admin only; omitted from company-safe profile. }
        position: { type: string }
        profile_photo_url: { type: string, nullable: true }
        skills:
          type: array
          items: { $ref: '#/components/schemas/Skill' }
        languages:
          type: array
          items: { type: string }
        documents:
          type: array
          items: { $ref: '#/components/schemas/WorkerDocumentMetadata' }
        work_history_summary: { type: string, nullable: true }
        work_history:
          type: array
          items: { $ref: '#/components/schemas/WorkExperience' }
        gender:
          type: string
          nullable: true
          enum: [male, female]
        whatsapp_available: { type: boolean, example: true }
        status: { $ref: '#/components/schemas/WorkerStatus' }
        reject_reason: { type: string, nullable: true }
        availability: { type: boolean }
        worker_class: { $ref: '#/components/schemas/WorkerClass' }
        is_foc_training:
          type: boolean
          example: false
          description: Admin-only flag. Not exposed in company-safe worker profile responses.
        foc_training_note: { type: string, nullable: true }
        foc_training_updated_at: { type: string, format: date-time, nullable: true }
        foc_training_updated_by_id: { type: string, format: uuid, nullable: true }
        rating_avg: { type: number, example: 4.5 }
        rating_count: { type: integer, example: 12 }
        rating_summary:
          type: object
          properties:
            average: { type: number, example: 4.5 }
            count: { type: integer, example: 12 }
        created_at: { type: string, format: date-time }
    CompanyVisibleWorkerProfile:
      type: object
      description: Worker profile visible to a company. Private contact fields such as phone/email are intentionally omitted.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        position: { type: string, nullable: true }
        profile_photo_url: { type: string, nullable: true }
        skills:
          type: array
          items: { $ref: '#/components/schemas/Skill' }
        languages:
          type: array
          items: { type: string }
        documents:
          type: array
          description: Only company-visible documents are returned.
          items: { $ref: '#/components/schemas/WorkerDocumentMetadata' }
        work_history_summary: { type: string, nullable: true }
        work_history:
          type: array
          items: { $ref: '#/components/schemas/WorkExperience' }
        rating_avg: { type: number, example: 4.5 }
        rating_count: { type: integer, example: 12 }
...
    RatingSummary:
      type: object
      properties:
        rating_avg: { type: number, example: 4.8 }
        rating_count: { type: integer, example: 9 }
        data:
          type: array
          items: { $ref: '#/components/schemas/Rating' }
...
    RejectRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, example: Documents are incomplete. }
...
    UpdateWorkerFocTrainingRequest:
      type: object
      required: [worker_ids, is_foc_training]
      properties:
        worker_ids:
          type: array
          minItems: 1
          maxItems: 200
          items: { type: string, format: uuid }
        is_foc_training:
          type: boolean
          description: true adds selected workers to F.O.C. training; false removes them.
        note:
          type: string
          nullable: true
          maxLength: 1000
          example: Training cost covered by SET Service.
...
paths:
...
  /workers/me:
    get:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      summary: Get approved worker profile
      responses:
        '200': { description: Worker profile, content: { application/json: { schema: { $ref: '#/components/schemas/WorkerProfile' } } } }
        '403': { description: Worker is not approved }
    patch:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      summary: Update approved worker profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skills: { type: array, items: { $ref: '#/components/schemas/Skill' } }
                languages: { type: array, items: { type: string } }
                email: { type: string, format: email, nullable: true, description: Optional worker email. Must be unique when provided. }
                work_history_summary: { type: string, nullable: true, maxLength: 2000 }
                work_history:
                  type: array
                  items: { $ref: '#/components/schemas/WorkExperience' }
                gender:
                  type: string
                  nullable: true
                  enum: [male, female]
                whatsapp_available: { type: boolean }
                availability: { type: boolean }
      responses:
        '200': { description: Updated worker profile }
  /workers/me/profile-photo:
    post:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      summary: Upload worker profile photo
      description: Approved workers can upload JPEG, PNG, or WebP profile photos. Local storage is dev-only; production should use S3/R2.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201': { description: Updated worker profile, content: { application/json: { schema: { $ref: '#/components/schemas/WorkerProfile' } } } }
        '400': { description: Invalid upload file, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Worker is not approved }
  /workers/me/documents:
    post:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      summary: Upload worker document
      description: Approved workers can upload PDF or image documents. Content is validated by extension, MIME, magic bytes, an image/PDF parser, and a malware scanner before promotion from quarantine to ready storage. `criminal_record` is private to worker/admin by default; `health_certificate` is company-visible.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, file]
              properties:
                type:
                  type: string
                  enum: [health_certificate, criminal_record]
                  example: health_certificate
                file:
                  type: string
                  format: binary
      responses:
        '201': { description: Updated worker profile, content: { application/json: { schema: { $ref: '#/components/schemas/WorkerProfile' } } } }
        '400': { description: Invalid upload file or document type, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Worker is not approved }
        '413': { description: Upload exceeds the five MiB limit, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '415': { description: Extension, declared MIME, or magic bytes are not allowed or do not match, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Corrupt, active-content, polyglot, or malware-positive file, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: Required malware scanner is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /workers/me/documents/{type}:
    delete:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      operationId: deleteMyWorkerDocument
      summary: Delete one document owned by the authenticated worker
      description: The document metadata is replaced with a non-downloadable soft-delete tombstone and an audit event in one transaction. Only a worker can delete their own document. The owned private object is removed after commit on a best-effort basis; foreign or malformed keys are never deleted.
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string, enum: [health_certificate, criminal_record] }
      responses:
        '200':
          description: Document soft-deleted
          headers:
            X-Request-ID: { $ref: '#/components/headers/XRequestId' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerDocumentDeletionResponse' }
        '401': { description: Bearer token is missing or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Authenticated account is not an approved worker, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Worker-owned document does not exist or was already deleted, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /workers/me/account-deletion-request:
    post:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      operationId: requestMyWorkerAccountDeletion
      summary: Request deletion of the authenticated worker account
      description: Immediately soft-deletes and anonymizes the worker-owned profile and authentication data, revokes sessions and device tokens, tombstones documents, and writes an append-only audit event in one transaction. Historical operational relations remain intact; no database row is hard-deleted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerAccountDeletionRequest' }
      responses:
        '202':
          description: Account deletion request accepted and applied
          headers:
            X-Request-ID: { $ref: '#/components/headers/XRequestId' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerAccountDeletionResponse' }
        '400': { description: Explicit confirmation is missing or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Bearer token is missing or invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Authenticated account is not a worker, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Worker account does not exist, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /private-worker-documents/{token}:
    get:
      tags: [Workers]
      security: []
      operationId: downloadLocalPrivateWorkerDocument
      summary: Read a local-development private document capability URL
      description: |
        Local storage only. The opaque, short-lived signed token is the capability; responses are
        never cached and the token must not be logged, persisted, or forwarded. Production object
        storage returns its own short-lived signed URL instead.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Private document bytes
          headers:
            X-Request-ID: { $ref: '#/components/headers/XRequestId' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
            Cache-Control: { schema: { type: string }, description: "private, no-store, max-age=0" }
            Pragma: { schema: { type: string, enum: [no-cache] } }
            Content-Disposition: { schema: { type: string }, description: Forces download as an attachment. }
            X-Content-Type-Options: { schema: { type: string, enum: [nosniff] } }
            Content-Security-Policy: { schema: { type: string }, description: Blocks active document content in the browser. }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { description: Token is invalid, expired, or the private object does not exist }
  /workers/{id}/documents/{type}/download:
    get:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      operationId: authorizeWorkerDocumentDownload
      summary: Authorize access and issue a short-lived private document URL
      description: |
        Worker ownership, related approved company visibility, or admin `view_workers` permission is
        required on every authorization request. Only ready documents with a clean scan result are
        eligible. The returned provider capability expires after at most fifteen minutes, is audit
        logged, and must not be logged, persisted, placed in analytics, or sent as a referrer.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: type
          required: true
          schema: { type: string, enum: [health_certificate, criminal_record] }
      responses:
        '200':
          description: Short-lived signed download URL
          headers:
            X-Request-ID: { $ref: '#/components/headers/XRequestId' }
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
            Cache-Control: { schema: { type: string }, description: "private, no-store, max-age=0" }
            Pragma: { schema: { type: string, enum: [no-cache] } }
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_in_seconds, expires_at]
                properties:
                  url: { type: string }
                  expires_in_seconds: { type: integer, minimum: 1, maximum: 900, example: 300 }
                  expires_at: { type: string, format: date-time }
        '401': { description: Bearer token is missing or invalid }
        '403': { description: Caller is not allowed to access this document }
        '404': { description: Worker or document not found }
        '410': { description: Document is rejected, deleted, legacy-public, foreign, or not clean/ready }
  /workers/{id}/company-profile:
    get:
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      summary: Get company-safe worker profile
      description: Approved company can view a safe worker profile only when the worker is assigned to one of its orders. Phone/email/private contact data is not returned.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Company-safe worker profile, content: { application/json: { schema: { $ref: '#/components/schemas/CompanyVisibleWorkerProfile' } } } }
        '403': { description: Company is not approved or cannot view this worker }
        '404': { description: Worker not found for this company }
...
  /admin/workers:
    get:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: List workers with pagination, search, status filter, and created_at sort
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/WorkerStatus' }
        - name: worker_class
          in: query
          schema: { $ref: '#/components/schemas/WorkerClass' }
        - name: foc_training
          in: query
          schema: { type: string, enum: [foc, non_foc] }
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200': { description: Paginated workers }
        '403': { description: Requires super_admin or admin with view_workers }
  /admin/workers/foc-training:
    patch:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: Add or remove selected workers from F.O.C. training
      description: Requires super_admin or admin with manage_workers. Creates worker_foc_training_added or worker_foc_training_removed audit log entries. This admin-only metadata is not exposed to company-safe worker profile responses.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWorkerFocTrainingRequest' }
            examples:
              add:
                value:
                  worker_ids: ["{{worker_id_1}}", "{{worker_id_2}}"]
                  is_foc_training: true
                  note: "Təlim xərci SET Service tərəfindən qarşılanır."
              remove:
                value:
                  worker_ids: ["{{worker_id_1}}"]
                  is_foc_training: false
      responses:
        '200': { description: F.O.C. training status updated }
        '403': { description: Requires super_admin or admin with manage_workers, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: One or more workers were not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/workers/{id}:
    get:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: Get full worker details, documents, OTP status, and approval status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Worker detail }
        '404': { description: Worker not found }
  /admin/workers/{id}/approve:
    patch:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: Approve worker, create audit log, and notify worker
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Worker approved }
  /admin/workers/{id}/reject:
    patch:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: Reject worker with reason, create audit log, and notify worker
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectRequest' }
      responses:
        '200': { description: Worker rejected }
  /admin/workers/{id}/class:
    patch:
      tags: [Admin - Workers]
      security: [{ BearerAuth: [] }]
      summary: Assign or clear worker class
      description: Requires super_admin or admin with manage_workers. Classifies an approved or pending worker as A, B, or C. The value is admin-only and is not exposed to company-safe worker profile responses.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_class]
              properties:
                worker_class: { $ref: '#/components/schemas/WorkerClass' }
            examples:
              setClass:
                value: { worker_class: A }
              clearClass:
                value: { worker_class: null }
      responses:
        '200': { description: Worker class updated, content: { application/json: { schema: { $ref: '#/components/schemas/WorkerProfile' } } } }
        '400': { description: Invalid class, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Requires super_admin or admin with manage_workers, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Worker not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /workers/{id}/ratings:
    get:
      tags: [Ratings]
      security: [{ BearerAuth: [] }]
      summary: Get worker rating history (admin canonical route)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Worker ratings, content: { application/json: { schema: { $ref: '#/components/schemas/RatingSummary' } } } }
        '403': { description: Requires super_admin or admin with view_workers, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Worker not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...